Refactor HTTP API
